refactor(@dpc-sdp/ripple-tide-landing-page): :recycle: use composable for webform submit
